refactor: split AddWordScreen composables into sibling files
AddWordScreen.kt held 11 composables in 791 lines and carried a
@file:Suppress("TooManyFunctions") to satisfy detekt's file-level
function count. Splits it along the pattern already used by
AddWordDialogs.kt and AddWordScreenPreviews.kt in this package:

- AddWordScreen.kt keeps AddWordScreen and AddWordContent
- AddWordInputSections.kt: WordInputCard, AiToggleRow,
  TranslationInputCard, BidirectionalOptionSection, TranslationDirectionRow
- AddWordFeedbackSections.kt: ActionButtonsRow, SuccessOverlayCard,
  ErrorMessageCard, PendingWordsSection

Moved composables go from private to internal since they're now
cross-file within the same package; no call sites or behavior change.
