+++ app/src/main/java/com/procrastilearn/app/ui/screens/AddWordScreen.kt
@@ -1,44 +1,23 @@
-@file:Suppress("TooManyFunctions")
-
 package com.procrastilearn.app.ui.screens
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.heightIn
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.List
-import androidx.compose.material.icons.automirrored.filled.ArrowForward
-import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.Delete
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.Checkbox
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.OutlinedTextFieldDefaults
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
@@ -47,17 +26,15 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.input.KeyboardCapitalization
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.zIndex
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.procrastilearn.app.R
+import com.procrastilearn.app.domain.model.AiTranslationDirection
 import com.procrastilearn.app.ui.AddWordLoadingAction
 import com.procrastilearn.app.ui.AddWordPreviewContent
 import com.procrastilearn.app.ui.AddWordViewModel
 import com.procrastilearn.app.ui.PendingWordUi
-import com.procrastilearn.app.domain.model.AiTranslationDirection
 import kotlinx.coroutines.delay
 
 @Suppress("MagicNumber")
@@ -322,469 +299,5 @@ internal fun AddWordContent(
                 onCancel = onExistingWordDialogCancel,
             )
         }
-    }
-}
-
-@Composable
-private fun WordInputCard(
-    word: String,
-    wordError: String?,
-    isLoading: Boolean,
-    openAiAvailable: Boolean,
-    useAiForTranslation: Boolean,
-    translationDirection: AiTranslationDirection,
-    nativeLanguageCode: String,
-    targetLanguageCode: String,
-    onWordChange: (String) -> Unit,
-    onUseAiToggle: (Boolean) -> Unit,
-    onTranslationDirectionToggle: () -> Unit,
-) {
-    Card(
-        modifier = Modifier.fillMaxWidth(),
-        colors =
-            CardDefaults.cardColors(
-                containerColor = MaterialTheme.colorScheme.surface,
-            ),
-        elevation = CardDefaults.cardElevation(defaultElevation = 2.dp),
-    ) {
-        Column(
-            modifier =
-                Modifier
-                    .fillMaxWidth()
-                    .padding(16.dp),
-        ) {
-            if (openAiAvailable) {
-                AiToggleRow(
-                    useAiForTranslation = useAiForTranslation,
-                    translationDirection = translationDirection,
-                    nativeLanguageCode = nativeLanguageCode,
-                    targetLanguageCode = targetLanguageCode,
-                    onUseAiToggle = onUseAiToggle,
-                    onTranslationDirectionToggle = onTranslationDirectionToggle,
-                )
-            }
-            val disableWordInput = isLoading && openAiAvailable && useAiForTranslation
-            OutlinedTextField(
-                value = word,
-                onValueChange = onWordChange,
-                label = { Text(stringResource(R.string.add_word_label_word)) },
-                placeholder = { Text(stringResource(R.string.add_word_placeholder_word)) },
-                isError = wordError != null,
-                supportingText = wordError?.let { { Text(it) } },
-                modifier = Modifier.fillMaxWidth(),
-                singleLine = true,
-                enabled = !disableWordInput,
-                readOnly = disableWordInput,
-                colors =
-                    OutlinedTextFieldDefaults.colors(
-                        focusedBorderColor = MaterialTheme.colorScheme.primary,
-                        unfocusedBorderColor = MaterialTheme.colorScheme.outline,
-                    ),
-            )
-        }
-    }
-}
-
-@Composable
-private fun AiToggleRow(
-    useAiForTranslation: Boolean,
-    translationDirection: AiTranslationDirection,
-    nativeLanguageCode: String,
-    targetLanguageCode: String,
-    onUseAiToggle: (Boolean) -> Unit,
-    onTranslationDirectionToggle: () -> Unit,
-) {
-    if (useAiForTranslation) {
-        TranslationDirectionRow(
-            direction = translationDirection,
-            nativeLanguageCode = nativeLanguageCode,
-            targetLanguageCode = targetLanguageCode,
-            onToggle = onTranslationDirectionToggle,
-            modifier = Modifier.fillMaxWidth(),
-        )
-        Spacer(modifier = Modifier.height(8.dp))
-    }
-    Row(modifier = Modifier.fillMaxWidth()) {
-        Checkbox(
-            checked = useAiForTranslation,
-            onCheckedChange = { onUseAiToggle(it) },
-        )
-        Spacer(modifier = Modifier.width(8.dp))
-        Text(
-            text = stringResource(R.string.add_word_use_ai_toggle),
-            style = MaterialTheme.typography.bodyMedium,
-            color = MaterialTheme.colorScheme.onSurface,
-            modifier = Modifier.align(Alignment.CenterVertically),
-        )
-    }
-    Spacer(modifier = Modifier.height(8.dp))
-}
-
-@Composable
-private fun TranslationInputCard(
-    translation: String,
-    translationError: String?,
-    useAiForTranslation: Boolean,
-    openAiAvailable: Boolean,
-    onTranslationChange: (String) -> Unit,
-) {
-    // Hidden when AI is used and key is present
-    if (useAiForTranslation && openAiAvailable) return
-
-    Card(
-        modifier = Modifier.fillMaxWidth(),
-        colors =
-            CardDefaults.cardColors(
-                containerColor = MaterialTheme.colorScheme.surface,
-            ),
-        elevation = CardDefaults.cardElevation(defaultElevation = 2.dp),
-    ) {
-        Column(
-            modifier =
-                Modifier
-                    .fillMaxWidth()
-                    .padding(16.dp),
-        ) {
-            OutlinedTextField(
-                value = translation,
-                onValueChange = onTranslationChange,
-                label = { Text(stringResource(R.string.add_word_label_translation)) },
-                placeholder = { Text(stringResource(R.string.add_word_placeholder_translation)) },
-                isError = translationError != null,
-                supportingText = translationError?.let { { Text(it) } },
-                modifier =
-                    Modifier
-                        .fillMaxWidth()
-                        .heightIn(min = 120.dp, max = 240.dp),
-                singleLine = false,
-                minLines = 4,
-                maxLines = 8,
-                keyboardOptions =
-                    KeyboardOptions(
-                        capitalization = KeyboardCapitalization.Sentences,
-                    ),
-                colors =
-                    OutlinedTextFieldDefaults.colors(
-                        focusedBorderColor = MaterialTheme.colorScheme.primary,
-                        unfocusedBorderColor = MaterialTheme.colorScheme.outline,
-                    ),
-            )
-        }
-    }
-}
-
-@Composable
-private fun BidirectionalOptionSection(
-    visible: Boolean,
-    bidirectional: Boolean,
-    isCustomizing: Boolean,
-    backwardPromptOverride: String,
-    backwardAnswerOverride: String,
-    onBidirectionalToggle: (Boolean) -> Unit,
-    onCustomizeToggle: () -> Unit,
-    onBackwardPromptOverrideChange: (String) -> Unit,
-    onBackwardAnswerOverrideChange: (String) -> Unit,
-) {
-    if (!visible) return
-
-    Card(
-        modifier = Modifier.fillMaxWidth(),
-        colors =
-            CardDefaults.cardColors(
-                containerColor = MaterialTheme.colorScheme.surface,
-            ),
-        elevation = CardDefaults.cardElevation(defaultElevation = 2.dp),
-    ) {
-        Column(modifier = Modifier.fillMaxWidth().padding(16.dp)) {
-            Row(modifier = Modifier.fillMaxWidth()) {
-                Checkbox(
-                    checked = bidirectional,
-                    onCheckedChange = onBidirectionalToggle,
-                )
-                Spacer(modifier = Modifier.width(8.dp))
-                Text(
-                    text = stringResource(R.string.add_word_bidirectional_toggle),
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = MaterialTheme.colorScheme.onSurface,
-                    modifier = Modifier.align(Alignment.CenterVertically),
-                )
-            }
-
-            if (bidirectional) {
-                TextButton(onClick = onCustomizeToggle) {
-                    Text(
-                        text =
-                            stringResource(
-                                if (isCustomizing) {
-                                    R.string.add_word_customize_backward_hide
-                                } else {
-                                    R.string.add_word_customize_backward_show
-                                },
-                            ),
-                    )
-                }
-
-                AnimatedVisibility(visible = isCustomizing) {
-                    Column {
-                        OutlinedTextField(
-                            value = backwardPromptOverride,
-                            onValueChange = onBackwardPromptOverrideChange,
-                            label = { Text(stringResource(R.string.add_word_backward_prompt_label)) },
-                            placeholder = { Text(stringResource(R.string.add_word_backward_prompt_placeholder)) },
-                            modifier = Modifier.fillMaxWidth(),
-                        )
-                        Spacer(modifier = Modifier.height(8.dp))
-                        OutlinedTextField(
-                            value = backwardAnswerOverride,
-                            onValueChange = onBackwardAnswerOverrideChange,
-                            label = { Text(stringResource(R.string.add_word_backward_answer_label)) },
-                            placeholder = { Text(stringResource(R.string.add_word_backward_answer_placeholder)) },
-                            modifier = Modifier.fillMaxWidth(),
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun ActionButtonsRow(
-    openAiAvailable: Boolean,
-    useAiForTranslation: Boolean,
-    isLoading: Boolean,
-    isOnline: Boolean,
-    loadingAction: AddWordLoadingAction?,
-    isAddLaterMode: Boolean,
-    onPreviewClick: () -> Unit,
-    onAddClick: () -> Unit,
-) {
-    val showPreviewButton = openAiAvailable && useAiForTranslation
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        horizontalArrangement = if (showPreviewButton) Arrangement.spacedBy(12.dp) else Arrangement.Start,
-    ) {
-        if (showPreviewButton) {
-            Button(
-                onClick = onPreviewClick,
-                modifier =
-                    Modifier
-                        .weight(1f)
-                        .height(56.dp),
-                enabled = !isLoading && isOnline,
-                colors =
-                    ButtonDefaults.buttonColors(
-                        containerColor = MaterialTheme.colorScheme.secondaryContainer,
-                        contentColor = MaterialTheme.colorScheme.onSecondaryContainer,
-                    ),
-                elevation =
-                    ButtonDefaults.buttonElevation(
-                        defaultElevation = 2.dp,
-                        pressedElevation = 4.dp,
-                    ),
-            ) {
-                if (isLoading && loadingAction == AddWordLoadingAction.PREVIEW) {
-                    CircularProgressIndicator(
-                        modifier = Modifier.size(20.dp),
-                        color = MaterialTheme.colorScheme.onSecondaryContainer,
-                    )
-                } else {
-                    Text(
-                        text = stringResource(R.string.add_word_button_preview),
-                        style = MaterialTheme.typography.titleMedium,
-                    )
-                }
-            }
-        }
-
-        val addModifier =
-            if (showPreviewButton) {
-                Modifier
-                    .weight(1f)
-                    .height(56.dp)
-            } else {
-                Modifier
-                    .fillMaxWidth()
-                    .height(56.dp)
-            }
-
-        Button(
-            onClick = onAddClick,
-            modifier = addModifier,
-            enabled = !isLoading,
-            colors =
-                ButtonDefaults.buttonColors(
-                    containerColor = MaterialTheme.colorScheme.primary,
-                ),
-            elevation =
-                ButtonDefaults.buttonElevation(
-                    defaultElevation = 4.dp,
-                    pressedElevation = 8.dp,
-                ),
-        ) {
-            if (isLoading && loadingAction == AddWordLoadingAction.ADD) {
-                CircularProgressIndicator(
-                    modifier = Modifier.size(24.dp),
-                    color = MaterialTheme.colorScheme.onPrimary,
-                )
-            } else {
-                Icon(
-                    imageVector = Icons.Default.Add,
-                    contentDescription = stringResource(R.string.action_add),
-                    modifier = Modifier.size(24.dp),
-                )
-                Spacer(modifier = Modifier.width(8.dp))
-                Text(
-                    text =
-                        stringResource(
-                            if (isAddLaterMode) {
-                                R.string.add_word_button_add_later
-                            } else {
-                                R.string.action_add
-                            },
-                        ),
-                    style = MaterialTheme.typography.titleMedium,
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun SuccessOverlayCard(successMessage: String?) {
-    Card(
-        colors =
-            CardDefaults.cardColors(
-                containerColor = MaterialTheme.colorScheme.primaryContainer,
-            ),
-        elevation = CardDefaults.cardElevation(defaultElevation = 8.dp),
-    ) {
-        Column(
-            modifier = Modifier.padding(32.dp),
-            horizontalAlignment = Alignment.CenterHorizontally,
-        ) {
-            Text(
-                text = "✓",
-                style = MaterialTheme.typography.displayMedium,
-                color = MaterialTheme.colorScheme.primary,
-            )
-            Spacer(modifier = Modifier.height(8.dp))
-            Text(
-                text = successMessage ?: stringResource(R.string.add_word_success_default),
-                style = MaterialTheme.typography.titleMedium,
-                color = MaterialTheme.colorScheme.onPrimaryContainer,
-            )
-        }
-    }
-}
-
-@Composable
-private fun ErrorMessageCard(errorMessage: String?) {
-    Card(
-        modifier =
-            Modifier
-                .fillMaxWidth()
-                .padding(top = 16.dp),
-        colors =
-            CardDefaults.cardColors(
-                containerColor = MaterialTheme.colorScheme.errorContainer,
-            ),
-    ) {
-        Text(
-            text = errorMessage ?: "",
-            modifier = Modifier.padding(16.dp),
-            color = MaterialTheme.colorScheme.onErrorContainer,
-            textAlign = TextAlign.Center,
-        )
-    }
-}
-
-@Composable
-internal fun PendingWordsSection(
-    pendingWords: List<PendingWordUi>,
-    onDeletePendingWord: (Long) -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    Column(modifier = modifier) {
-        Text(
-            text = stringResource(R.string.add_word_pending_title),
-            style = MaterialTheme.typography.titleSmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
-        )
-        Spacer(modifier = Modifier.height(8.dp))
-        Card(
-            modifier = Modifier.fillMaxWidth(),
-            colors =
-                CardDefaults.cardColors(
-                    containerColor = MaterialTheme.colorScheme.surface,
-                ),
-            elevation = CardDefaults.cardElevation(defaultElevation = 2.dp),
-        ) {
-            Column(modifier = Modifier.fillMaxWidth()) {
-                pendingWords.forEachIndexed { index, pendingWord ->
-                    Row(
-                        modifier =
-                            Modifier
-                                .fillMaxWidth()
-                                .padding(horizontal = 16.dp, vertical = 4.dp),
-                        verticalAlignment = Alignment.CenterVertically,
-                    ) {
-                        Text(
-                            text = pendingWord.word,
-                            style = MaterialTheme.typography.bodyLarge,
-                            modifier = Modifier.weight(1f),
-                        )
-                        IconButton(onClick = { onDeletePendingWord(pendingWord.id) }) {
-                            Icon(
-                                imageVector = Icons.Default.Delete,
-                                contentDescription = stringResource(R.string.add_word_pending_delete),
-                                tint = MaterialTheme.colorScheme.error,
-                            )
-                        }
-                    }
-                    if (index != pendingWords.lastIndex) {
-                        HorizontalDivider()
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-internal fun TranslationDirectionRow(
-    direction: AiTranslationDirection,
-    nativeLanguageCode: String,
-    targetLanguageCode: String,
-    onToggle: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    val isNativeToTarget = direction == AiTranslationDirection.NATIVE_TO_TARGET
-    val startLabel = if (isNativeToTarget) nativeLanguageCode else targetLanguageCode
-    val endLabel = if (isNativeToTarget) targetLanguageCode else nativeLanguageCode
-    Row(
-        modifier = modifier,
-        horizontalArrangement = Arrangement.Center,
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        Text(
-            text = startLabel,
-            style = MaterialTheme.typography.labelLarge,
-            color = MaterialTheme.colorScheme.onSurface,
-        )
-        Spacer(modifier = Modifier.width(8.dp))
-        IconButton(onClick = onToggle) {
-            Icon(
-                imageVector = Icons.AutoMirrored.Default.ArrowForward,
-                contentDescription = stringResource(R.string.add_word_toggle_direction),
-                tint = MaterialTheme.colorScheme.primary,
-            )
-        }
-        Spacer(modifier = Modifier.width(8.dp))
-        Text(
-            text = endLabel,
-            style = MaterialTheme.typography.labelLarge,
-            color = MaterialTheme.colorScheme.onSurface,
-        )
     }
 }
